refactor: split/move 'ui' to submodule
