+++ setup.py
@@ -42,13 +42,14 @@ setup(
     packages=[
         "vunit",
         "vunit.com",
-        "vunit.test",
         "vunit.parsing",
         "vunit.parsing.verilog",
-        "vunit.vivado",
+        "vunit.test",
         "vunit.test.lint",
         "vunit.test.unit",
         "vunit.test.acceptance",
+        "vunit.ui",
+        "vunit.vivado",
     ],
     package_data={"vunit": DATA_FILES},
     zip_safe=False,
